--- pkg/controller/compliancesuite/compliancesuite_suite_test.go
@@ -1,13 +1,0 @@
-package compliancesuite_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestComplianceSuite(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "ComplianceSuite Suite")
-}
